+++ ToDoList/Model/BasicWorkerRepresentable.swift
@@ -2,7 +2,7 @@ import Foundation
 
 public protocol BasicWorkerRepresentable: EmployeeRepresentable {
     var breakHours: TimePeriod { get }
-    var deskNumber: Int64 { get }
+    var deskNumber: Int { get }
     var name: String { get }
     var salary: Double { get }
 }
@@ -10,9 +10,9 @@ public protocol BasicWorkerRepresentable: EmployeeRepresentable {
 // Class is dependent on EmployeeConstructor - dependency injection is not possible. Revisit this implementation.
 class BasicWorkerConstructor: EmployeeConstructor, BasicWorkerRepresentable {
     let breakHours: TimePeriod
-    let deskNumber: Int64
+    let deskNumber: Int
 
-    init(name: String, salary: Double, breakHours: TimePeriod, deskNumber: Int64) {
+    init(name: String, salary: Double, breakHours: TimePeriod, deskNumber: Int) {
         self.breakHours = breakHours
         self.deskNumber = deskNumber
         super.init(name: name, salary: salary, employeeType: .basicWorker)
